+++ packages/utils/src/arr/includes.ts
@@ -1,4 +1,4 @@
-import type { Suggest, UnknownArray } from '@ylfjuk-core/types';
+import type { Suggest, UnknownArray } from "@ylfjuk-core/types";
 
 /**
  * Checks if an array includes a specific element.
@@ -6,11 +6,9 @@ import type { Suggest, UnknownArray } from '@ylfjuk-core/types';
  * @note A wrapper around {@link Array.prototype.includes} that allows to check against any type
  *
  * @since 0.0.8
+ * @modified 0.0.10
  */
-export const arrIncludes = <
-    const Arr extends UnknownArray,
-    const El extends Suggest<Arr[number]>
->(
-    arr: Arr,
-    el: El
+export const includes = <const Arr extends UnknownArray, const El extends Suggest<Arr[number]>>(
+	arr: Arr,
+	el: El,
 ): boolean => arr.includes(el);