+++ packages/utils/src/json/index.ts
@@ -1,16 +1,13 @@
-import { jsonParse } from "./parse";
-import { jsonSafeParse } from "./safe-parse";
-import { jsonStringify } from "./stringify";
-
 /**
- * Utility functions for {@link JSON}.
+ * Utility module for {@link JSON}.
  *
  * @since 0.0.9
+ * @modified 0.0.10
  */
-export const Json = {
-	parse: jsonParse,
-	safeParse: jsonSafeParse,
-    stringify: jsonStringify,
-};
+export * as Json from "./Json";
 
-export { jsonParse, jsonSafeParse, jsonStringify };
+export {
+	parse as jsonParse,
+	safeParse as jsonSafeParse,
+	stringify as jsonStringify,
+} from "./Json";