+++ packages/utils/src/obj/keys.ts
@@ -4,6 +4,7 @@
  * @note A wrapper around {@link Object.keys} that retains the type of the keys.
  *
  * @since 0.0.8
+ * @modified 0.0.10
  */
-export const keysOf = <const T extends object>(obj: T): (keyof T)[] =>
-    Object.keys(obj) as (keyof T)[];
+export const keys = <const T extends object>(obj: T): (keyof T)[] =>
+	Object.keys(obj) as (keyof T)[];