+++ packages/utils/src/str/Str.ts
@@ -1,0 +1,2 @@
+export * from "./ends-with";
+export * from "./starts-with";